initial refactoring
* refactored main classes into lines and regions2d directories
