+++ echoregions/__init__.py
@@ -1,6 +1,7 @@
-from .convert.api import read_evl, read_evr
-from .convert.utils import parse_simrad_fname_time, parse_time
-from .formats import Lines, Regions2D
+from .utils.api import read_evl, read_evr
+from .utils.utils import parse_simrad_fname_time, parse_time
+from .lines.lines import Lines
+from .regions2d.regions2d import Regions2D
 
 __all__ = [
     "read_evl",
