--- echoregions/formats/__init__.py
@@ -1,5 +1,0 @@
-from .geometry import Geometry
-from .lines import Lines
-from .regions2d import Regions2D
-
-__all__ = ["Geometry", "Lines", "Regions2D"]
