+++ echoregions/regions2d/regions2d_mask.py
@@ -1,5 +1,4 @@
 from typing import Any, List, NewType, Union
-
 import matplotlib
 import numpy as np
 import regionmask
