+++ echoregions/tests/test_line.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-
 import pandas as pd
 import pytest
